Minor Tweaks and Wallet Creation
